Refactor code to folllow a MVC structure
